refactor mesh code into MeshController
